// packages/conv-ai-provider/src/types.ts
export type MessageId = ReturnType<typeof uuidv4>;

// A user prompt message and its resulting AI response message share a request ID
// (This informs us which AI response corresponds to which user prompt)
export type RequestId = ReturnType<typeof uuidv4>;

// schema of actions from websocket notification streaming message chunks
export type WSNotificationBotMessageClientAction = BotMessageClientAction & {
  arguments: {
    text: {
      streamId: string;
    };
  };
};

export enum ConnectionState {
  Connected = 'Connected',
  Reconnecting = `Reconnecting`,
  Closed = `Closed`,
}

export enum MessageRequestType {
  RegenerationRequest = 'RegenerationRequest',
  UserPrompt = 'UserPrompt',
}

type ISODateTimeString = string; // e.g. "2024-05-03T17:32:19.318Z"

export type ConversationRateLimitError =
  | 'UserRateLimitExceeded'
  | 'ServiceRateLimitExceeded'
  | 'TokenLimitExceeded';

export type ConversationError =
  | ConversationRateLimitError
  | 'ContentFilterFailed'
  | 'InternalError';

type MessageEvent = {
  startStreamingResponse: (
    content: string,
    messageId: MessageId,
    requestId: RequestId,
    actions: WSNotificationBotMessageClientAction[],
  ) => void;
  finishStreamingResponse: (requestId: RequestId) => void;
  updateStreamingResponse: (content: string, messageId: MessageId) => void;
  errorStreamingResponse: (conversationError: ConversationError, messageId: MessageId) => void;
  connectionConnected: () => void;
  connectionReconnecting: () => void;
  connectionClose: () => void;
};

export type Emitter = TypedEventEmitter<MessageEvent>;
// ...
export enum ConvApiV2ActionType {
  DisplayRichText = 'DisplayRichText',
  DisplayError = 'DisplayError',
}

// ref: https://github.rbx.com/Roblox/grasshopper/blob/master/packages/clients/schemas/convAiPlatformOrchestration/v1.json
export type BotMessageClientAction = {
  actionTime: ISODateTimeString;
  actionSummary: string; // autosummarized by assistant
  actionType: 'client_action';
  arguments: {
    text: string; // assistant message text
    errorId?: string; // error ID if actionType is DisplayError
  };
  command: ConvApiV2ActionType;
  plugin: ''; // empty string for docs assistant
};